changed conflicting data types

+++ ft_strncpy.c
@@ -6,15 +6,15 @@
 /*   By: tchewa <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 10:49:19 by tchewa            #+#    #+#             */
-/*   Updated: 2019/05/24 11:12:02 by tchewa           ###   ########.fr       */
+/*   Updated: 2019/05/30 10:53:09 by tchewa           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strncpy(char *dest, char *src, unsigned int n)
+char	*ft_strncpy(char *dest, char *src, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
